Working on some HTTP actions with ASINetworking
Pretty happy with how easy this was to implement. ASINetworking maintains its
own cookie vault, so there's no need to manually grab the cookies from the
authentication request and include them in the later requests to open doors.

The hard part about fetching out the nateberg-berg arguments (ref, branch) is
that it needs to be done probably from the command line, and I don't think that
NSTask is a very good solution. Unfortunately, ~10m of googling hasn't turned up
any better alternatives, so it's probably worth a shot.